requirements.txt
watchdog==3.0.0
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
supabase==1.2.0 